requirements.txt
lxml==4.9.1
python-telegram-bot==13.13
PyYAML==6.0
requests==2.28.1
